+++ src/main/java/co/edu/uniquindio/bookyourstay/modelo/factory/Apartamento.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public class Apartamento extends Alojamiento {
 
-    public Apartamento(String id, String nombre, String ciudad, String descripcion, float precioNoche, byte capacidadMaxima, Image imagenAlojamiento, List<String> servicios) {
+    public Apartamento(String id, String nombre, String ciudad, String descripcion, float precioNoche, int capacidadMaxima, Image imagenAlojamiento, List<String> servicios) {
         super(id, nombre, ciudad, descripcion, precioNoche, capacidadMaxima, imagenAlojamiento, servicios);
     }
